Enum steam_vdf::ValveData

pub enum ValveData {
    List(OsString, Vec<ValveData>),
    String(OsString, OsString),
    Bytes4(OsString, [u8; 4]),
    EndOfList,
}

Represents data in a vdf file. All data except the list terminator has a property name and property value.

Variants
List(OsString, Vec<ValveData>)

A list. Begins with 0x00, ends with EndOfList, or 0x08 Internal format: 0x00 string name list_contents EndOfList

String(OsString, OsString)

A string. Begins with 0x01. All standalone strings end with 0x00. The string ValveData is simply two strings. Internal format: 0x01 string name string value

Bytes4(OsString, [u8; 4])

4 bytes. Begins with 0x02. Internal format: 0x02 string name 4 bytes

EndOfList

Only considered a type for the use of parsing lists. This is always the last element in the list. Begins with 0x08. Internal format: 0x08

Methods

impl ValveData

fn data_type(&self) -> ValveDataType

Convert from data to type.