Function steam_vdf::write_data

pub fn write_data<T: Write>(output: &mut T, data: &ValveData) -> Result<()>

Writes to the given output the given ValveData. * If it succeeds without errors, it returns Ok(()) * If it has errors, it returns Err with an io error.

This does not write the extra 0x08 at the end of shortcuts.vdf. In order to produce a valid file, you will need to write 0x08 when you're done.

Open the file you are working with in a hex editor and check if this will produce proper output. If the output is not proper, it may be discarded, and data may be lost.